\docType{package}
\name{types}
\alias{types}
\alias{types-package}
\title{Static types for R}
\description{
types provides a simple type annotation for R that is usable in scripts,
in the R console and in packages. It is intended as a convention to allow other
packages to use the type information to provide error checking,
automatic documentation or optimizations.

The annotations are syntactically valid R code rather than comments, which
provides additional assurance they are specified properly. However this package
does not do anything with the type annotations, they have no effect on the
calculated result.

`?` when used on the command line continues to work as before, accessing the R
help pages for the topic requested.
}
